# Divide large constant buffer into subsets and implement push constants for Vulkan performance