--- a/neo/shaders/renderParmSet1.inc.hlsl
+++ b/neo/shaders/renderParmSet1.inc.hlsl
@@ -39,8 +39,6 @@ struct renderParmSet1_t
 	float4 rpMVPmatrixZ;
 	float4 rpMVPmatrixW;
 	
-	float4 rpProjectionMatrixZ;
-
 	float4 rpOverbright;
 };