This PR depends on https://github.com/RobertBeckebans/nvrhi/pull/6

This PR also solves #763 for Apple Silicon performance and rendering artifact elimination, as well as #804 for Intel Integrated GPU support with Vulkan (slow, but works).

This is a fairly large pull request that implements the following:

1. Separates the single large constant buffer into renderparm subsets (12 in total: 3 of 128 bytes in size, 6 greater than 128 bytes but less than 256 bytes, and 3 greater than 256 bytes but less than 1024 bytes)
2. Adds new binding layout types to associate and differentiate between the new subsets (`BINDING_LAYOUT_GBUFFER, BINDING_LAYOUT_GBUFFER_SKINNED, BINDING_LAYOUT_TEXTURE, BINDING_LAYOUT_TEXTURE_SKINNED, BINDING_LAYOUT_WOBBLESKY, BINDING_LAYOUT_SSGI, BINDING_LAYOUT_SSGI_SKINNED, BINDING_LAYOUT_POST_PROCESS`)
3. Implements push constants for Vulkan and DX12 across all platforms: Linux, macOS, Windows.  This has varying degrees of performance improvement, the largest being on Vulkan for Linux and macOS.  Windows Vulkan shows modest improvement dependant on the GPU vendor (Nvidia's 256 byte limit is better than AMD's 128 byte limit on Windows).  Windows DX12 shows no performance improvement when using push constants vs. volatile constant buffers.  Given this I have defined a new boolean `r_useDX12PushConstants` cvar which is turned off by default.  This can optionally be turned on using autoexec.cfg for experimentation.
4. Reduced the volatile constant max buffer count from 16,384 to 8,192.  I believe this is sufficient but if testing reveals differently, then it could be boosted back up.  Note that when push constants are enabled it reduces the requirement.
5. Adds basic infrastructure for static constant buffers but these are disabled for now.  This could be a possibility for the future but further subset refactoring would likely be needed, and sync issues would have to be resolved.
6. Simplified Vulkan code by removing the barrier command list - not needed if command submit is done in correct order
7. Added CPU and GPU usage % counters to the on-screen HUD display for all platforms
8. Added MoltenVK's Vulkan-to-Metal encoding time to the HUD when available for macOS only.  This capability will be available in the next release of MoltenVK (1.2.6)
9. Addressed a nasty bug with mis-calibrated GPU timers for macOS caused by a regression in the MoltenVK 1.2.5 release. Use either MoltenVK release <= 1.2.4 or the coming 1.2.6 release with this pull request to avoid the issue.
10. Modified the **ssao_compute** fix to support multiple HLSL versions, not just 2021.  I found that my changes were required for compiling DXIL for D3D12 while remaining compatible with SPIRV for Vulkan, otherwise I would get compile failures on Windows DX12 (`select` does not seem to be supported when compiling for DXIL).
11. A minor CMakeLists fix primarily for Xcode that cleans up precompiled.h-xxxxx.gch.tmp files left around when the ZERO_CHECK target runs for regeneration.
12. **UPDATED**: Tested on Apple Silicon (M1) and disabled GPU Skinning for macOS arm64 to eliminate rendering artifacts.  I will have to look at this later to see what is going on with GPU skinning on Apple Silicon.
13. **UPDATED**: Modified `cmake-macos-*.sh` and `cmake-xcode-*.sh` build scripts for openal-soft path portability across x86 and Apple Silicon.  Thanks to @asemarafa for the code.
14. **UPDATED**: Works around missing Vulkan `shaderStorageImageReadWithoutFormat` device feature on Intel GPUs, and individually activates `VK_KHR_fragment_shading_rate` sub-features vs. all or none (this is supported by nvrhi).
15. **FINAL UPDATE**: Fixed uniforms change detection logic (orthogonal to push constants) which has a very significant positive impact on performance.  See updated performance timings below.  Also added new cvar `r_useVulkanPushConstants` (default on) which is useful for performance comparisons.

Tested on Windows 10 (AMD and Nvidia), Linux Manjaro, and macOS Ventura 13.5

Performance timings for this PR vs. current master, generated using a simple home-made timedemo:

**Windows Nvidia System (1070 Ti)**
DX12: 263 fps before, ~~255~~ **360 fps after** (with r_useDX12PushConstants = 0) --> significant improvement
Vulkan: 218 fps before, ~~233~~ **333 fps after** --> significant improvement

**Windows AMD System (6600 XT)**
DX12: 295 fps before, ~~285~~ **305 fps after** (with r_useDX12PushConstants = 0) --> neutral/positive improvement
Vulkan: 150 fps before, ~~155~~ **160 fps after** --> neutral/positive improvement

**Linux AMD System (6600 XT)**
Vulkan: 150 fps before, ~~210~~ **270 fps after** --> large improvement

**macOS AMD System (6600 XT)**
Vulkan: 77 fps before, ~~177~~ **245 fps after** --> very large improvement

**macOS Apple Silicon System (M1 Air)** (**_new_**)
Vulkan: 6 fps before, **85 fps after** --> massive improvement

Some possible explanations are as follows:

1. Windows nvrhi DX12 push constants are implemented as D3D12 root constants which are just a special constant buffer of limited size (about 240 bytes for RBDoom3BFG).  I suspect there is no real performance gain by using these over nvrhi's volatile constant buffers which appear to be quite efficient with low synchronization overhead on Windows.  As a result `r_useDX12PushConstants = 0` by default.
2. Nvidia offers 256 bytes of Vulkan push constants on Windows, while AMD offers only 128 bytes of Vulkan push constants on Windows.  For this reason AMD Vulkan users on Windows will likely see a small positive change in performance.  Nvidia users on Windows will see a greater ~~but still modest~~ improvement for Vulkan.
3. On Linux both Nvidia and AMD can offer 256 bytes of Vulkan push constants. Perhaps AMD's driver implementation on Linux favours push constants due to reduced synchronization overhead and the performance gains are quite significant.  I would be interested in seeing some perf timings for Nvidia on Linux.
4. On macOS/MoltenVK/Metal there are 4096 bytes of push constants.  This means that virtually all of the draw transactions are handled with push constants and volatile constant buffers are not needed.  This is a good thing since constant buffer synchronization on macOS/MoltenVK/Metal appears to be expensive.  Push constants do not incur any synchronization overheads and are delivered with command submission.
5. Overall I think this is a useful change, especially for Vulkan users on Linux and macOS.  I would go as far as saying it is mandatory for macOS on Apple Silicon, as the game is virtually unplayable without it. The Windows benefits are ~~less, but still in the positive direction for Vulkan~~ significant for DX12 & Vulkan on Nvidia, but less but so on AMD,  at least for my test systems/GPUs.

@RobertBeckebans here is the spreadsheet I used to create the binding layout type to shader/renderparm subset mapping:
[Binding to Shader Mapping v4.xlsx](https://github.com/RobertBeckebans/RBDOOM-3-BFG/files/12877986/Binding.to.Shader.Mapping.v4.xlsx)