@RobertBeckebans 2023-10-18
The results are awesome but I need quite a lot of testing and time to review this code. I aim for a new public release in December before Christmas.