@SRSaunders 2023-10-18
Thanks @RobertBeckebans - take your time on review.  It took me a while to figure this out anyways, plus a bunch of work with the MoltenVK project to diagnose and optimize performance issues.  It was my summer/fall background project :)

If you have questions please don't hesitate to post a message or send an email.  I will do my best to answer.