@SRSaunders 2023-10-21
Hi @RobertBeckebans.  I have pushed my final design update to this branch.  Commit https://github.com/RobertBeckebans/RBDOOM-3-BFG/pull/818/commits/2063c72912a76f07f545dddaa806e0e7869d0b1c adds one small but critical change for performance - fixing the uniforms change detection logic which reduces how often volatile constant buffers are written for the larger renderparm sets.  This has a huge effect which is orthogonal to push constants, and has a positive impact across all platforms: Windows, Linux, & macOS.  See updated timing info above.  I don't quite know why my AMD 6600XT GPU seems to be limited when running Windows/Vulkan, with other OSs surpassing it with the same card.  Perhaps something about my windows drivers or installation - I am not sure.