@SRSaunders 2023-11-30
I added a few minor things due to dependencies on previous changes within this PR:

1. Added a few additional comments to explain some of the important changes in this PR
2. Don't allocate constant buffers unless required (i.e. when push constants disabled for binding layout type)
3. For the statistics overlay HUD, I added GPU Memory usage and smoothing to CPU/GPU usage percentages
4. Made the VMA header file visible within the IDE source tree under libs/vma (CMakeLists change)