@SRSaunders 2023-12-20
Oops, just realized that ImmediateMode was broken for using debug tools with push constants enabled.

Fixed in https://github.com/RobertBeckebans/RBDOOM-3-BFG/pull/818/commits/ee3b6f9b2f0754cf0bcaf47a2d124f2c337defe1