@SRSaunders 2023-12-28
**macOS only**: I just added support for the new `VK_EXT_layer_settings` extension used for configuring the upcoming **MoltenVK 1.2.7 / Vulkan SDK 1.3.272.0**. See commit https://github.com/RobertBeckebans/RBDOOM-3-BFG/pull/818/commits/f3c65eef3e75ea2846cffd97aff1fd2f9bf92f7f.

@RobertBeckebans The last several commits have been added mostly due to dependencies on earlier changes within this PR.  Unless something comes up in your testing, I don't plan any more changes here.