@labrnth 2024-01-01
Also noticed some strange artifacts on the weapons (not seeing this elsewhere). See attached screenshot.
![rbdoom-3-bfg-20240101-111545-001](https://github.com/RobertBeckebans/RBDOOM-3-BFG/assets/13229225/65b55e88-f122-4a1e-b2e9-0d38f501f62a)

I have observed a massive increase in FPS though, so thats great.