@SRSaunders 2024-01-01
@labrnth Thanks for trying out this PR.  The compile issue you found was fixed in master with commit ab663a76.  Perhaps you synced before this was merged recently.

Regarding the weapons artifacts, are you running on an Intel or Apple Silicon Mac?  I thought this problem was only visible on Apple Silicon - where I disabled GPU skinning to fix it.  If you are running on Intel that is quite interesting and may indicate a more general problem with skinning.  In this case you could try setting `r_useGPUSkinning 0` in your console, or adding `set r_useGPUSkinning 0` to your autoexec.cfg file.  Please let me know your results.

Good news on the performance improvement.  That was my main goal for this effort.