@SRSaunders 2024-01-02
Thanks for catching the `r_useGPUSkinning` reset in multiplayer.  Solved in https://github.com/RobertBeckebans/RBDOOM-3-BFG/pull/818/commits/3b6598b88d71bc443c431a04e0f075410679ea36 (this cvar is not a cheat in any case).