@SRSaunders 2024-01-21
@RobertBeckebans I would like to know what you are planning with this PR.  I have a few other improvements for Vulkan that I would like to submit (e.g. use Vulkan dynamic functions vs. static linkage for VMA config, Optick, and MoltenVK), but I don't want to keep adding on here, especially if you are not planning to merge in the near term.

I could simplify things by dividing this into two PRs that could be treated independently:

1. Separate the Uniforms Subsetting, Push Constants and Uniforms Change Detection improvements which is focussed solely on performance for both Vulkan and DX12.  And yes, DX12 benefits from uniforms subsetting with the new change detection logic even without push constants enabled.
2. Separate the HUD GPU Memory usage & CPU/GPU Usage % features (DX12 and Vulkan), generic Vulkan changes (incl. Intel iGPU fixes, dynamic functions, etc), and macOS/MoltenVK-only stuff.

I would then close this large PR and you could look at two new smaller PRs independently.

Would this help?  If so would you look at either of the new PRs for your upcoming release?